front: move occupancy blocks in a separate file
